LayoutTests/http/tests/notifications/icon-exists-show-alert-during-load.html

<!DOCTYPE html>
<html>
<head>
    <script type="text/javascript">
        function log(message)
        {
            document.getElementById("result").innerHTML += message + "<br>";
        }
        
        function runTests()
        {
            if (window.layoutTestController) {
                layoutTestController.grantDesktopNotificationPermission("http://127.0.0.1");
                layoutTestController.dumpAsText();
                layoutTestController.waitUntilDone();
            }

            if (!window.webkitNotifications) {
                log("FAIL: No webkitNotifications interface!");
            }
            var N = window.webkitNotifications.createNotification("resources/icon-exists.png", "New E-mail", "Meet me tonight at 8!");
            N.ondisplay = function() { log("PASS: display event invoked."); layoutTestController.notifyDone(); }
            N.show();
            alert("Load should be deferred due to alert");
        }
    </script>
</head>
<body>
<p>Sending notification, and then alert. Load should be deferred but continue after the alert.</p>
<div id="result"></div>    
    
<script type="text/javascript">
runTests();
</script>
</body>
</html>
